review request: viewModel separation
